+++ pom.xml
@@ -24,7 +24,7 @@
                     <archive>
                         <manifest>
                             <addClasspath>true</addClasspath>
-                            <mainClass>com.x_tornado10.Main</mainClass>
+                            <mainClass>com.x_tornado10.lccp.LCCP</mainClass>
                         </manifest>
                     </archive>
                     <descriptorRefs>
