--- src/main/java/com/x_tornado10/Events/EventListener.java
@@ -1,7 +1,0 @@
-package com.x_tornado10.Events;
-
-import com.x_tornado10.Events.Events.Event;
-
-public interface EventListener {
-    void onEvent(Event event);
-}
