+++ src/main/java/com/x_tornado10/lccp/Events/EventListener.java
@@ -1,0 +1,7 @@
+package com.x_tornado10.lccp.Events;
+
+import com.x_tornado10.lccp.Events.Events.Event;
+
+public interface EventListener {
+    void onEvent(Event event);
+}
