+++ src/main/java/com/x_tornado10/lccp/Events/Events/SaveEvent.java
@@ -1,4 +1,4 @@
-package com.x_tornado10.Events.Events;
+package com.x_tornado10.lccp.Events.Events;
 
 public class SaveEvent extends Event{
     @Override
