+++ src/main/java/com/x_tornado10/lccp/Logger/Log_Level.java
@@ -1,4 +1,4 @@
-package com.x_tornado10.Logger;
+package com.x_tornado10.lccp.Logger;
 
 public interface Log_Level {
     boolean isEnabled();
