+++ src/main/java/com/x_tornado10/lccp/util/Networking.java
@@ -1,6 +1,6 @@
-package com.x_tornado10.util;
+package com.x_tornado10.lccp.util;
 
-import com.x_tornado10.Main;
+import com.x_tornado10.lccp.LCCP;
 
 import java.io.*;
 import java.net.Socket;
@@ -28,47 +28,47 @@ public final class Networking {
 
             File fileToSend = new File(fileToSendPath);
 
-            Main.logger.debug("Received request to send '" + fileToSend.getAbsolutePath() +"' to " + serverIP4 + ":" + serverPort + "!");
-            Main.logger.debug("Inspecting file...");
-            Main.logger.debug("File name: " + fileToSend.getName());
-            Main.logger.debug("File size: " + ((fileToSend.length() / 1024) / 1024) + "MB");
-            Main.logger.debug("File type: " + fileToSend.getName().split("\\.")[1].toUpperCase());
-            Main.logger.info("Sending File: '" + fileToSend.getAbsolutePath() + "' to " + serverIP4 + ":" + serverPort);
+            LCCP.logger.debug("Received request to send '" + fileToSend.getAbsolutePath() +"' to " + serverIP4 + ":" + serverPort + "!");
+            LCCP.logger.debug("Inspecting file...");
+            LCCP.logger.debug("File name: " + fileToSend.getName());
+            LCCP.logger.debug("File size: " + ((fileToSend.length() / 1024) / 1024) + "MB");
+            LCCP.logger.debug("File type: " + fileToSend.getName().split("\\.")[1].toUpperCase());
+            LCCP.logger.info("Sending File: '" + fileToSend.getAbsolutePath() + "' to " + serverIP4 + ":" + serverPort);
 
 
             try {
-                Main.logger.debug("Opening new socket: " + serverIP4 + ":" + serverPort);
+                LCCP.logger.debug("Opening new socket: " + serverIP4 + ":" + serverPort);
                 // opening new socket
                 Socket socket = new Socket(serverIP4, serverPort);
-                Main.logger.debug("Successfully established connection!");
+                LCCP.logger.debug("Successfully established connection!");
 
-                Main.logger.debug("Opening output streams...");
+                LCCP.logger.debug("Opening output streams...");
                 // opening the output streams used to transfer data
                 OutputStream outputStream = socket.getOutputStream();
-                Main.logger.debug("Successfully opened output streams!");
+                LCCP.logger.debug("Successfully opened output streams!");
 
                 // sending file metadata
-                Main.logger.debug("Sending file metadata...");
-                Main.logger.debug("Sending file name...");
+                LCCP.logger.debug("Sending file metadata...");
+                LCCP.logger.debug("Sending file name...");
                 // sending file name
-                Main.logger.fatal(fileToSend.getName());
+                LCCP.logger.fatal(fileToSend.getName());
                 outputStream.write(fileToSend.getName().strip().getBytes());
-                Main.logger.debug("Sending file size...");
+                LCCP.logger.debug("Sending file size...");
                 // sending file size in bytes
-                Main.logger.fatal(String.valueOf(fileToSend.length()));
+                LCCP.logger.fatal(String.valueOf(fileToSend.length()));
                 outputStream.write(String.valueOf(fileToSend.length()).getBytes());
                 outputStream.flush();
-                Main.logger.debug("Successfully send file metadata!");
+                LCCP.logger.debug("Successfully send file metadata!");
 
-                Main.logger.debug("Opening new FileInputStream to read the main file content...");
+                LCCP.logger.debug("Opening new FileInputStream to read the main file content...");
                 // opening file input stream to read main file content
                 FileInputStream fileInputStream = new FileInputStream(fileToSend);
                 BufferedInputStream bufferedInputStream = new BufferedInputStream(fileInputStream);
                 // get socket output stream
-                Main.logger.debug("Defining new 8129B buffer...");
+                LCCP.logger.debug("Defining new 8129B buffer...");
                 // defining new 8KB buffer
                 byte[] buffer = new byte[8192];
-                Main.logger.debug("Successfully created new BufferedInputStream for the FileInputStream!");
+                LCCP.logger.debug("Successfully created new BufferedInputStream for the FileInputStream!");
                 int count;
 
                 long fileSize = fileToSend.length();
@@ -82,7 +82,7 @@ public final class Networking {
                 long delay = 1000;
                 long lastDisplay = System.currentTimeMillis() - 1000;
 
-                Main.logger.debug("Sending main file contents...");
+                LCCP.logger.debug("Sending main file contents...");
                 // reading and sending main file content using buffer
                 while ((count = bufferedInputStream.read(buffer)) > 0) {
                     if (System.currentTimeMillis() - lastDisplay >= delay) {
@@ -116,7 +116,7 @@ public final class Networking {
                         }
                         result.append(parts[2]).append("s");
 
-                        Main.logger.debug("Transferring File: " + mbTransferredSize + "MB / " + mbFileSize + "MB -- " + (double) Math.round(percent * 1000) / 1000 + "% -- Speed: " + mbPerSecond + "MB/S -- ETA: " + result.toString().trim());
+                        LCCP.logger.debug("Transferring File: " + mbTransferredSize + "MB / " + mbFileSize + "MB -- " + (double) Math.round(percent * 1000) / 1000 + "% -- Speed: " + mbPerSecond + "MB/S -- ETA: " + result.toString().trim());
                         lastTransferredSize = transferredSize;
                     }
                     outputStream.write(buffer, 0, count);
@@ -125,20 +125,20 @@ public final class Networking {
                     mbTransferredSize = (double) Math.round(temp1 * 1000) / 1000;
                 }
                 outputStream.flush();
-                Main.logger.debug("Successfully send file contents!");
+                LCCP.logger.debug("Successfully send file contents!");
 
-                Main.logger.debug("Closing socket and streams...");
+                LCCP.logger.debug("Closing socket and streams...");
                 // closing socket and all streams to free up resources
                 bufferedInputStream.close();
                 outputStream.close();
                 socket.close();
-                Main.logger.debug("Successfully closed socket and streams!");
-                Main.logger.debug("Sending complete!");
+                LCCP.logger.debug("Successfully closed socket and streams!");
+                LCCP.logger.debug("Sending complete!");
             } catch (IOException e) {
                 e.printStackTrace();
                 return false;
             }
-            Main.logger.info("Successfully send file to server!");
+            LCCP.logger.info("Successfully send file to server!");
             return true;
         }
     }
