--- src/main/resources/Calculator.css
@@ -1,4 +1,0 @@
-entry {
-  color: black;
-  background-color: white;
-}
