+++ src/main/resources/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: com.x_tornado10.Main
+Main-Class: com.x_tornado10.lccp.LCCP
